requirements.txt
asgiref==3.7.2
Django==4.2.4
django-crispy-forms==2.0
pyperclip==1.8.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
django-import-export==3.3.0